// angular-sample/angular-webapp/client/app/document/models/document-category-enum.ts
export enum DocumentCategoryEnum {
        Policy = 64,
        Contract = 256,
        Handbook = 1024,
        Documents = 2048,
        SiteVisit = 4096,
        BusinessDocument = 8192,
        Uploads = 0,
        ELHandbook = 16384,
        Icons = 32,
        RiskAssessment = 512,
        SiteVisitQA = ELHandbook * 2,
        CheckList = SiteVisitQA * 2,
        PEL = CheckList + 1,
        HS = PEL + 1,
        PELCitation = HS + 1,
        HSCitation = PELCitation + 1,
        PELQualityMarkCitation = HSCitation + 1,
        HSQualityMarkCitation = PELQualityMarkCitation + 1,
        Appraisal = 9000,
        Disciplinary = Appraisal + 1,
        Emails = Disciplinary + 1,
        Grievance = Emails + 1,
        Leaver = Grievance + 1,
        NewStarter = Leaver + 1,
        ScannedDocument = NewStarter + 1,
        Other = ScannedDocument + 1,
        PersonalDocuments = Other + 1,
        ContractTemplate = 16,
        EmployeeContract = 128,
        StandardHazardIcons = PersonalDocuments + 1,
        StandardControlIcons = StandardHazardIcons + 1,
        Certificates = 65543,
        MethodStatements = 513,
        Task = SiteVisit + 1,
        ComplianceCertificates = 65545,
        AccidentLogs = 51200,
        ConstructionPhasePlans = 5120,
        FireRiskAssessment = 514,
        DBS = 9011,
        CarePolicies = 515,
        General = 65572,
        COSHH = 516,
        CompanyPolicies = 517,
        Performance = 90909
}
export enum DocumentChangesEnum {
        NoChange = 0,
        ContentChanged = 1,
        BlockAdded = ContentChanged * 2,
        BlocksReordered = BlockAdded * 2,
        BlockRemoved = BlocksReordered * 2,
        FieldsChanged = BlockRemoved * 2,
        TitleChanged = FieldsChanged * 2,
        FormatingChanged = TitleChanged * 2,
        BlockSwitchOn = FormatingChanged * 2,
        BlockSwitchOff = BlockSwitchOn * 2,
        Arhived = BlockSwitchOff * 2,
        OptionsChanges = Arhived * 2,
        OrderIndex = OptionsChanges * 2
}